feat: full feature sprint — leads, admin panel, mobile nav, emails
- Fix Gvamax price parsing (Argentine dot format) and price display (es-AR locale)
- Filter properties without title or images from listing
- Add image lightbox on property detail page
- Restructure property detail: title → photos → description
- Add leads system: localStorage persistence, admin panel view, Gvamax + email notification
- Add admin profile management: avatar upload, name/username, password change
- Add mobile bottom navigation bar for admin panel (5-tab fixed nav)
- Add Mi Cuenta section in admin: photo, personal data, password with validation
- Replace nodemailer with Resend for transactional emails
- Add /api/send-confirmation endpoint: user confirmation email after form submission
- Add /api/send-verification-code + /api/verify-code: 6-digit OTP for password recovery
- Add forgot password flow in Login (4 steps: email → code → new password → done)
- Add userStorageService and leadStorageService for localStorage persistence
- Add .env.example with Resend and DNS configuration instructions

+++ services/syncService.ts
@@ -56,7 +56,7 @@ const mapGvamaxToProperty = (item: any): Property => {
     description: cleanHtml(item.descripcion),
     type: mapPropertyType(item.tipoInmueble || ''),
     operation: mapOperationType(item.tipoOperacion || ''),
-    price: parseFloat(String(item.precio || '0').replace(/[^0-9.]/g, '')) || 0,
+    price: parseInt(String(item.precio || '0').replace(/[^0-9]/g, '')) || 0,
     currency: item.moneda === 'P' ? 'ARS' : 'USD',
     neighborhood: item.ubicacion?.barrio || '',
     location: {
